requirements.txt
Flask==3.0.0
flask-mysql-connector==1.1.0
mysql-connector-python==8.2.0
Werkzeug==3.0.1
python-dotenv==1.0.0
